+++ src/repositories/raw_history/Pacs.008.001.10.repository.ts
@@ -20,11 +20,11 @@ export const Pacs008Repo: CrudRepository<Pacs008> = {
       : { data: [], total: 0 };
   },
 
-  get: async function (id): Promise<Pacs008 | null> {
+  get: async function ({ id, tenantId }): Promise<Pacs008 | null> {
     const queryRes = await handleExecuteSqlStatement<{ document: Pacs008 }>(
       {
-        text: 'SELECT document FROM pacs008 WHERE messageid = $1;',
-        values: [id],
+        text: 'SELECT document FROM pacs008 WHERE messageid = $1 AND tenantid = $2;',
+        values: [id, tenantId],
       } satisfies PgQueryConfig,
       'raw_history',
     );
@@ -43,22 +43,22 @@ export const Pacs008Repo: CrudRepository<Pacs008> = {
     return queryRes.rows[0].document;
   },
 
-  update: async function (id: string, payload: Pacs008): Promise<Pacs008 | null> {
+  update: async function ({ id, tenantId }, payload: Pacs008): Promise<Pacs008 | null> {
     const queryRes = await handleExecuteSqlStatement<{ document: Pacs008 }>(
       {
-        text: 'UPDATE pacs008 SET document = $1 WHERE messageid = $2 RETURNING document;',
-        values: [payload, id],
+        text: 'UPDATE pacs008 SET document = $1 WHERE messageid = $2 AND tenantid = $3 RETURNING document;',
+        values: [payload, id, tenantId],
       } satisfies PgQueryConfig,
       'raw_history',
     );
     return queryRes.rowCount ? queryRes.rows[0].document : null;
   },
 
-  remove: async function (id: string): Promise<boolean> {
+  remove: async function ({ id, tenantId }): Promise<boolean> {
     const queryRes = await handleExecuteSqlStatement<{ document: Pacs008 }>(
       {
-        text: 'DELETE FROM pacs008 WHERE messageid = $1;',
-        values: [id],
+        text: 'DELETE FROM pacs008 WHERE messageid = $1 AND tenantid = $2;',
+        values: [id, tenantId],
       } satisfies PgQueryConfig,
       'raw_history',
     );
